# Remove IsAscii fast paths from Equals and Compare OrdinalIgnoreCase