--- a/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.cs
+++ b/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.cs
@@ -17,7 +17,7 @@ using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
 using System.Buffers;
-using System.Text;
+using Internal.Runtime.CompilerServices;
 
 namespace System.Globalization
 {
@@ -529,57 +529,103 @@ namespace System.Globalization
         {
             Debug.Assert(indexA + lengthA <= strA.Length);
             Debug.Assert(indexB + lengthB <= strB.Length);
-            return CompareOrdinalIgnoreCase(strA.AsSpan(indexA, lengthA), strB.AsSpan(indexB, lengthB));
+            return CompareOrdinalIgnoreCase(
+                ref Unsafe.Add(ref strA.GetRawStringData(), indexA),
+                lengthA,
+                ref Unsafe.Add(ref strB.GetRawStringData(), indexB),
+                lengthB);
         }
 
-        internal static unsafe int CompareOrdinalIgnoreCase(ReadOnlySpan<char> strA, ReadOnlySpan<char> strB)
+        internal static int CompareOrdinalIgnoreCase(ReadOnlySpan<char> strA, ReadOnlySpan<char> strB)
         {
-            int length = Math.Min(strA.Length, strB.Length);
+            return CompareOrdinalIgnoreCase(ref MemoryMarshal.GetReference(strA), strA.Length, ref MemoryMarshal.GetReference(strB), strB.Length);
+        }
+
+        internal static int CompareOrdinalIgnoreCase(string strA, string strB)
+        {
+            return CompareOrdinalIgnoreCase(ref strA.GetRawStringData(), strA.Length, ref strB.GetRawStringData(), strB.Length);
+        }
+
+        internal static int CompareOrdinalIgnoreCase(ref char strA, int lengthA, ref char strB, int lengthB)
+        {
+            int length = Math.Min(lengthA, lengthB);
             int range = length;
 
-            fixed (char* ap = &MemoryMarshal.GetReference(strA))
-            fixed (char* bp = &MemoryMarshal.GetReference(strB))
+            ref char charA = ref strA;
+            ref char charB = ref strB;
+
+            // in InvariantMode we support all range and not only the ascii characters.
+            char maxChar = (GlobalizationMode.Invariant ? (char)0xFFFF : (char)0x7F);
+
+            while (length != 0 && charA <= maxChar && charB <= maxChar)
             {
-                char* a = ap;
-                char* b = bp;
-
-                // in InvariantMode we support all range and not only the ascii characters.
-                char maxChar = (char) (GlobalizationMode.Invariant ? 0xFFFF : 0x7F);
-
-                while (length != 0 && (*a <= maxChar) && (*b <= maxChar))
+                // Ordinal equals or lowercase equals if the result ends up in the a-z range 
+                if (charA == charB ||
+                    ((charA | 0x20) == (charB | 0x20) &&
+                        (uint)((charA | 0x20) - 'a') <= (uint)('z' - 'a')))
                 {
-                    int charA = *a;
-                    int charB = *b;
+                    length--;
+                    charA = ref Unsafe.Add(ref charA, 1);
+                    charB = ref Unsafe.Add(ref charB, 1);
+                }
+                else
+                {
+                    int currentA = charA;
+                    int currentB = charB;
 
-                    if (charA == charB)
-                    {
-                        a++; b++;
-                        length--;
-                        continue;
-                    }
-
-                    // uppercase both chars - notice that we need just one compare per char
-                    if ((uint)(charA - 'a') <= 'z' - 'a') charA -= 0x20;
-                    if ((uint)(charB - 'a') <= 'z' - 'a') charB -= 0x20;
+                    // Uppercase both chars if needed
+                    if ((uint)(charA - 'a') <= 'z' - 'a')
+                        currentA -= 0x20;
+                    if ((uint)(charB - 'a') <= 'z' - 'a')
+                        currentB -= 0x20;
 
                     // Return the (case-insensitive) difference between them.
-                    if (charA != charB)
-                        return charA - charB;
-
-                    // Next char
-                    a++; b++;
-                    length--;
+                    return currentA - currentB;
                 }
-
-                if (length == 0)
-                    return strA.Length - strB.Length;
-
-                Debug.Assert(!GlobalizationMode.Invariant);
-
-                range -= length;
-
-                return CompareStringOrdinalIgnoreCase(a, strA.Length - range, b, strB.Length - range);
             }
+
+            if (length == 0)
+                return lengthA - lengthB;
+
+            Debug.Assert(!GlobalizationMode.Invariant);
+
+            range -= length;
+
+            return CompareStringOrdinalIgnoreCase(ref charA, lengthA - range, ref charB, lengthB - range);
+        }
+
+
+        internal static bool EqualsOrdinalIgnoreCase(ref char strA, ref char strB, int length)
+        {
+            ref char charA = ref strA;
+            ref char charB = ref strB;
+
+            // in InvariantMode we support all range and not only the ascii characters.
+            char maxChar = (GlobalizationMode.Invariant ? (char)0xFFFF : (char)0x7F);
+
+            while (length != 0 && charA <= maxChar && charB <= maxChar)
+            {
+                // Ordinal equals or lowercase equals if the result ends up in the a-z range 
+                if (charA == charB ||
+                    ((charA | 0x20) == (charB | 0x20) &&
+                        (uint)((charA | 0x20) - 'a') <= (uint)('z' - 'a')))
+                {
+                    length--;
+                    charA = ref Unsafe.Add(ref charA, 1);
+                    charB = ref Unsafe.Add(ref charB, 1);
+                }
+                else
+                {
+                    return false;
+                }
+            }
+
+            if (length == 0)
+                return true;
+
+            Debug.Assert(!GlobalizationMode.Invariant);
+
+            return CompareStringOrdinalIgnoreCase(ref charA, length, ref charB, length) == 0;
         }
 
         ////////////////////////////////////////////////////////////////////////