For consideration.

Contributes to https://github.com/dotnet/coreclr/issues/14320

IsAscii is optimized for one specific scenario so the removal of that code does see a 10% regression in equalling old strings. However it takes 11 EqualsLateExitBothCached compares to earn back the initial expense of the EqualsLateExitNoCached compare. So both strings need to be fairly long lived.

Equals OrdinalIgnoreCase

Method | Before | After | Improvement |
----------------------------------|----------:|----------:|----------:|
EqualsEarlyExitBothCached | 26.50 ns | 14.43 ns | 0.8x |
EqualsEarlyExitRightCached | 98.35 ns | 13.76 ns | 6.0x |
EqualsEarlyExitLeftCached | 101.01 ns | 13.92 ns | 6.2x |
EqualsEarlyExitNoCached | 170.91 ns | 13.83 ns | 11.3x |
EqualsLateExitBothCached | 101.63 ns | 113.29 ns | **-0.1x** |
EqualsLateExitRightCached | 174.00 ns | 116.34 ns | 0.5x |
EqualsLateExitLeftCached | 176.17 ns | 113.41 ns | 0.5x |
EqualsLateExitNoCached | 243.51 ns | 114.05 ns | 1.1x |
EqualsUnicodeEarlyExitBothCached |  29.70 ns | 13.88 ns | 1.1x |
EqualsUnicodeEarlyExitRightCached |  33.98 ns | 14.06 ns | 1.4x |
EqualsUnicodeEarlyExitLeftCached |  29.39 ns | 13.91 ns | 1.1x |
EqualsUnicodeEarlyExitNoCached |  35.25 ns | 14.84 ns | 1.3x |
EqualsUnicodeLateExitBothCached | 148.29 ns | 127.02 ns | 0.1x |
EqualsUnicodeLateExitRightCached | 224.69 ns | 127.14 ns | 0.7x |
EqualsUnicodeLateExitLeftCached | 148.28 ns | 126.70 ns | 0.2x |
EqualsUnicodeLateExitNoCached | 223.75 ns | 127.45 ns | 0.7x |

Compare OrdinalIgnoreCase

Method | Before | After | Improvement |
----------------------------------- |----------:|----------:|----------:|
CompareEarlyExitBothCached |  27.46 ns | 15.72 ns | 0.7x |
CompareEarlyExitRightCached |  99.46 ns | 15.84 ns | 5.2x |
CompareEarlyExitLeftCached |  99.28 ns | 15.16 ns | 5.5x |
CompareEarlyExitNoCached | 174.50 ns | 15.22 ns | 10.4x |
CompareLateExitBothCached | 134.67 ns | 109.96 ns | 0.2x |
CompareLateExitRightCached | 210.49 ns | 110.25 ns | 0.9x |
CompareLateExitLeftCached | 207.10 ns | 110.03 ns | 0.8x |
CompareLateExitNoCached | 280.76 ns | 110.62 ns | 1.5x |
CompareUnicodeEarlyExitBothCached |  28.87 ns | 15.39 ns | 0.8x |
CompareUnicodeEarlyExitRightCached |  32.94 ns | 15.40 ns | 1.1x |
CompareUnicodeEarlyExitLeftCached |  28.93 ns | 15.97 ns | 0.8x |
CompareUnicodeEarlyExitNoCached |  33.21 ns | 15.30 ns | 1.1x |
CompareUnicodeLateExitBothCached | 147.15 ns | 131.04 ns | 0.1x |
CompareUnicodeLateExitRightCached | 227.18 ns | 131.10 ns | 0.7x |
CompareUnicodeLateExitLeftCached | 147.45 ns | 130.74 ns | 0.1x |
CompareUnicodeLateExitNoCached | 229.50 ns | 130.99 ns | 0.7x |

Equals OrdinalIgnoreCase now has a dependency on environment variables that it didn't have before on ASCII values. Broke the circular path with a custom bool string compare. 10ns quicker but its only there for that reason. Any other risk factors here with startup code calling into OrdinalIgnoreCase?