@jkotas 2018-05-14
@dotnet-bot test Windows_NT x64 Checked Innerloop Build and Test please