@jkotas 2018-05-14
@dotnet-bot test Ubuntu x64 Checked corefx_baseline